escape data in calendar popover
