+++ templates/calendar/user.html.twig
@@ -120,11 +120,13 @@
 
         function renderEventPopoverContent(eventObj)
         {
+            const escaper = kimai.getPlugin('escape');
+
             return '<div class="calendar-entry">' +
                     '<ul>' +
-                        '<li>' + '{{ 'label.customer'|trans }}: ' + eventObj.customer + '</li>' +
-                        '<li>' + '{{ 'label.project'|trans }}: ' + eventObj.project + '</li>' +
-                        '<li>' + '{{ 'label.activity'|trans }}: ' + eventObj.activity + '</li>' +
+                        '<li>' + '{{ 'label.customer'|trans }}: ' + escaper.escapeForHtml(eventObj.customer) + '</li>' +
+                        '<li>' + '{{ 'label.project'|trans }}: ' + escaper.escapeForHtml(eventObj.project) + '</li>' +
+                        '<li>' + '{{ 'label.activity'|trans }}: ' + escaper.escapeForHtml(eventObj.activity) + '</li>' +
                     '</ul>' +
                     (eventObj.description !== null || eventObj.tags.length > 0 ? '<hr>' : '') +
                     (eventObj.description ? '<p>' + eventObj.description + '</p>' : '') +
